turned on more objc warnings so i had to fix stuff

+++ Categories/Foundation/LjsWebCategories.m
@@ -46,7 +46,7 @@ static const int ddLogLevel = LOG_LEVEL_WARN;
   NSMutableString * output = [NSMutableString string];
   const unsigned char * source = (const unsigned char *)[self UTF8String];
   unsigned long sourceLen = strlen((const char *)source);
-  for (int i = 0; i < sourceLen; ++i) {
+  for (unsigned long i = 0; i < sourceLen; ++i) {
     const unsigned char thisChar = source[i];
     if (thisChar == ' '){
       [output appendString:@"+"];
